refactor(daemon): give shim connection types a home and dedupe helpers
`ShimConnection` and `PodIdentityVerifier` lived in listener.ts, so every
consumer — the k8s driver, channels, session, the dialer — imported a shared
contract from the legacy dial-in endpoint. The credential TTL constant plus
`withoutToken()` and `sanitizeWorkspaceRoot()` were duplicated verbatim in
listener.ts and dialer.ts.

Move all five into a new `src/shim/connection.ts`, keeping one copy of each
helper, and repoint every importer. No re-export is left in listener.ts, so
the dependency edge is now on the shared module rather than on the listener.
`shim/protocol.ts` stays a pure zod wire layer with no node imports.

Behavior is unchanged; this is a move plus an import repoint.
