+++ packages/daemon/test/cluster-acp-e2e.test.ts
@@ -4,7 +4,7 @@ import { fileURLToPath } from 'node:url'
 import { Backoff, FakeClock } from '@agentconnect.md/connection'
 import { AcpHost } from '../src/acp/acp-host.js'
 import { K8sDriver } from '../src/k8s/driver.js'
-import type { ShimConnection } from '../src/shim/listener.js'
+import type { ShimConnection } from '../src/shim/connection.js'
 import { ShimClient, type ShimTransport } from '../src/shim/client.js'
 import { ShimDialer } from '../src/shim/dialer.js'
 import { ShimServer } from '../src/shim/server.js'
