+++ packages/daemon/test/shim-cancellation.test.ts
@@ -5,7 +5,8 @@ import { mkdtempSync, rmSync, writeFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { Backoff, ClientTransport, FakeClock } from '@agentconnect.md/connection'
-import { ShimListener, type ShimConnection } from '../src/shim/listener.js'
+import type { ShimConnection } from '../src/shim/connection.js'
+import { ShimListener } from '../src/shim/listener.js'
 import { ShimClient, type ShimTransport } from '../src/shim/client.js'
 import { ShimChannel, ShimRequestAbortedError } from '../src/shim/channels.js'
 import { createExecHandler } from '../src/shim/exec-handler.js'
